moved prompt.py to ragas/llms
